LayoutTests/http/tests/security/xss-DENIED-getSVGDocument-object-expected.txt

CONSOLE MESSAGE: Unsafe JavaScript attempt to access frame with URL http://localhost:8080/security/resources/empty-svg.php from frame with URL http://127.0.0.1:8000/security/xss-DENIED-getSVGDocument-object.html. Domains, protocols and ports must match.


PASS
